Generate i32 const version of the tests
