+++ test/WasmSpec/baselines/chakra_i32.baseline
@@ -1,0 +1,1 @@
+975/975 tests passed.
