+++ test/WasmSpec/convert-test-suite/generateTests/index.js
@@ -2,9 +2,11 @@
 // Copyright (C) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 //-------------------------------------------------------------------------------------------------------
+const chakraIntConst = require("./chakra_intConst");
 
 const genTests = [
-  "chakra_i64"
+  {name: "chakra_i64", getContent: chakraIntConst.bind(null, "i64")},
+  {name: "chakra_i32", getContent: chakraIntConst.bind(null, "i32")},
 ];
 
-module.exports = genTests.map(f => ({name: f, getContent: require(`./${f}`)}));
+module.exports = genTests;
