+++ test/WasmSpec/rlexe.xml
@@ -20,6 +20,22 @@
   <test>
     <default>
       <files>spec.js</files>
+      <baseline>baselines/chakra_i32.baseline</baseline>
+      <compile-flags>-wasm -args testsuite/chakra/chakra_i32.wast -endargs</compile-flags>
+      <tags>exclude_xplat</tags>
+    </default>
+  </test>
+  <test>
+    <default>
+      <files>spec.js</files>
+      <baseline>baselines/chakra_i32.baseline</baseline>
+      <compile-flags>-wasm -args testsuite/chakra/chakra_i32.wast -endargs -nonative</compile-flags>
+      <tags>exclude_dynapogo,exclude_xplat</tags>
+    </default>
+  </test>
+  <test>
+    <default>
+      <files>spec.js</files>
       <baseline>baselines/address.baseline</baseline>
       <compile-flags>-wasm -args testsuite/core/address.wast -endargs</compile-flags>
       <tags>exclude_xplat</tags>
